backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
